Another try at Garrisons.  Ran into unique id issues.
Encountered an issue where I was not able to get unique unit Ids for the spawned Garrisons.
The place where I needed to get the unique Id was in TGO and importing Game would
cause a circular dependency.  I wonder if conforming to the original Layout
design more would resolve this.